\name{cleandata}
\alias{cleandata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function for filtering and trimming raw Licor data. 
}
\description{
This function will take a raw Licor data csv file and remove extra information in the file above the logged data, remove unnecessary columns for other functions in the package, and remove any logged data that is an error due to machine matching.  
}
\usage{
cleandata(data, csvname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
A data frame generated by using \code{read.csv} to import a spreadsheet that is generated by a LI-6800.
}
  \item{csvname}{
A file name indicating what the csv file produced by the function will be called. Name must be placed in quotes. Ex. "mycsvfile.csv". Manualy set location of csv file or it will be placed in the current working directory.
}
}
\value{
\code{cleandata} {Returns a csv spreadsheet that contains data from the data argument that was filtered for any outliers and data collection errors.  Extra headers and variables have also been removed. This cleaned spreadsheet will be used in \code{graphPack} and \code{analysis}.
}
}
\author{
Robert Twohey III
}
\examples{
#import example data
data3 <- system.file("extdata", "func1_example.csv", package = "licordataanalysis")

#run function
data3 <- read.csv(data3, header = FALSE)
cleaned <- cleandata(data3,"")
}

