\name{graphPack}
\alias{graphPack}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function for producing stepwise graphs of stomatal conductance (gsw) and net photosynthesis (A)
from cleaned Licor data
}
\description{
This function graphs the values of gsw and A at varrying CO2 concentrations (400ppm, 800 ppm,
100 ppm).
}
\usage{
graphPack(cleandata)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{cleandata}{
A data frame generated by using \code{read.csv} to import the csv produced by the function
\code{cleandata}. 
}
}

\value{
\item{gswgraph_gsw}{ A scatter plot with time, in seconds, on the x-axis.  The y-axis shows stomatal
conductance of the measured plant.  From 0-1680 seconds data is collected at 400ppm CO2, 800ppm from
1740-3480 seconds, and 100ppm from 3540 - end of x-axis.
}
\item{agraph_anet}{ A scatter plot with time, in seconds, on the x-axis.  The y-axis shows net
photosynthesis of the measured plant.  From 0-1680 seconds data is collected at 400ppm CO2, 800ppm
from 1740-3480 seconds, and 100ppm from 3540 - end of x-axis.
}
}
\author{
Robert Twohey III
}

\examples{
#import example data
data4 <- system.file("extdata", "func2_example.csv", package = "licordataanalysis")


#run function
data4 <- read.csv(data4)
graphs <- graphPack(data4)

graphs[[1]]
graphs[[2]]
}
